Renaming files and removing empty ones
