--- src/components/characterCard/CharacterCard.spec.js
@@ -1,1 +1,0 @@
-test("", () => {})
