--- src/components/gameCreation/GameCreation.spec.js
@@ -1,13 +1,0 @@
-import React from "react";
-import { render, cleanup } from "@testing-library/react";
-import GameCreation from "./gameCreation";
-
-afterEach(cleanup);
-
-describe("GameCreation tests", () => {
-	test("displays a gameCreation", () => {
-		// Arrange
-		// Act
-		// Assert
-	});
-});
