--- src/components/instructionsField/InstructionsField.spec.js
@@ -1,1 +1,0 @@
-test("", () => {})
